serverai_parafrasa/requirements.txt
flask
flask-cors
transformers
torch
python-docx
requests
sentencepiece
sacremoses
protobuf
tqdm
